+++ src/plugins/soundkonverter_codec_neroaac/soundkonverter_codec_neroaac.cpp
@@ -93,7 +93,7 @@ CodecWidget *soundkonverter_codec_neroaac::newCodecWidget()
     return qobject_cast<CodecWidget*>(widget);
 }
 
-unsigned int soundkonverter_codec_neroaac::convert( const KUrl& inputFile, const KUrl& outputFile, const QString& inputCodec, const QString& outputCodec, const ConversionOptions *_conversionOptions, TagData *tags, bool replayGain )
+int soundkonverter_codec_neroaac::convert( const KUrl& inputFile, const KUrl& outputFile, const QString& inputCodec, const QString& outputCodec, const ConversionOptions *_conversionOptions, TagData *tags, bool replayGain )
 {
     QStringList command = convertCommand( inputFile, outputFile, inputCodec, outputCodec, _conversionOptions, tags, replayGain );
     if( command.isEmpty() )
@@ -177,7 +177,7 @@ float soundkonverter_codec_neroaac::parseOutput( const QString& output, int leng
     QRegExp regEnc("Processed (\\d+) seconds");
     if( output.contains(regEnc) )
     {
-        return (float)regEnc.cap(1).toInt()*100/length;
+        return (float)regEnc.cap(1).toInt() * 100 / (float)length;
     }
 
     // no output when decoding
